Use toolforge envvars if available
If both of the relevant envvars are set (TOOL_REPLICA_USER +
TOOL_REPLICA_PASSWORD for the replicas, TOOL_TOOLSDB_USER +
TOOL_TOOLSDB_PASSWORD for toolsdb – they’re currently the same but might
diverge later), use them and skip reading the replica.my.cnf, which may
not be mounted. If only one envvar is set, don’t use it – this is mostly
“should never happen” territory but I think that makes slightly more
sense than trying to combine an incomplete set of envvars with the
replica.my.cnf file.

Bug: T339940